Create README.md
Write a GUI program that displays the front of a home design page. On the top of the front page displays a horizontal navigation. The user should be able to click on the command button, “Start Building Your Dream Space!”. After the user clicks the command, a pop-up window will ask the user to “Get Started”. The title of the interface declares “Welcome Home, [User’s Name]. Let’s Create an account.” This title will appear underneath the home design logo. The widgets included are two buttons: 
Displaying underneath name on center will be the “Get Started” button:
•	The user will have to enter
•	Name
•	Gender
•	Business name
•	Business email address
•	Business phone number 
•	A short bio:
o	Description of education and skills
o	Short personal info.
•	Add a profile image functionality.
•	Agree to the terms and conditions
•	“Design & Explore” button
•	One button underneath will show “Sign In” at the bottom of the “Create an Account” page.
•	Leave button
Else:
Displaying underneath the log in information will be a “Sign In” Button 
Another pop-up window will be displayed.
Two labels, one username (on top) and password (on bottom) will be shown in the center of the window.